Add limit filter to Database
This adds the limit Todos function to the database.  A `filterTodosByLimit()` method was made, which copys a portion of the arrray equal to the limit specified by the user.  This method is then utilized within the `getTodos()` method.  If  "limit" is contained within the query parameters, then `filterTodosByLimit()` will be called.