Sort Todos by Status
This adds a helper function that gives a integer value to the boolean values, so they can be compared for sorting alphabetically.  This also adds the sort by status function it self, which makes a call to `compareStatus()`.  This gives the user the ability to order todos by their status.